further namespace partitioning (Hash and DamExchange)

+++ Utilities/TemplateTricks.h
@@ -1,5 +1,4 @@
 #pragma once
-#include "IntegerTypes.h"
 
 //+------------------------------------------------------------------------------------------------------+
 //|      Trick by Andrei Alexandrescu, Dr. Dobb's, October 01, 2001                                      |
@@ -9,8 +8,7 @@
 //|      This is because different template instantiations are distinct types,                           |
 //|      so Int2Type<0> is different from Int2Type<1> and so on.                                         |
 //|                                                                                                      |
-//|      This trick is used in so-called "tag dispatching", a.k.a. partial specialization of             |
-//|      function templates. The universal move generator makes extensive use of this technique.         |
+//|      This trick is used in "tag dispatching", a.k.a. partial specialization of function templates.   |
 //+------------------------------------------------------------------------------------------------------+
 
-template<size_t> struct Int2Type {};
+template<int> struct Int2Type {};
